requirements.txt
streamlit
google-genai
python-dotenv
gTTS
google-generativeai
gtts
google-cloud-texttospeech
